Add WikidataIds decorator
For every link to a suitable page in the same wiki (ie not a Special page, or one in a different namespace)
look up its Wikidata ID, and attach that as an attribute of the link, suitable for the scraper to extract again.

wikidata-fetcher neatly encapsulates that lookup for us already, including
doing them in suitably sized batches.